alpha_vantage/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
requests==2.22.0
django-redis==5.0.0
celery==4.4.2
redis==3.5.3
